chore: 유효성 검사 함수 폴더 옮기고 적용
