+++ src/services/authClient.service.ts
@@ -3,6 +3,7 @@ import authClientRepository from "../repositories/authClient.repository";
 import { ILoginDataLocal, ISignUpDataLocal } from "../types";
 import { NotFoundError } from "../types/errors";
 import { filterSensitiveUserData, hashPassword, verifyPassword } from "../utils/auth.util";
+import { validateSignUpData } from "../utils/auth.util";
 import { generateAccessToken, generateRefreshToken } from "../utils/token.util";
 
 // ✅ 회원가입 - Local
@@ -11,6 +12,9 @@ async function create(
 ): Promise<Omit<ISignUpDataLocal, "hashedPassword" | "phone">> {
   // 비밀번호 해시
   const hashedPassword = await hashPassword(user.hashedPassword);
+
+  await validateSignUpData({ ...user, hashedPassword });
+
   const newClient = await authClientRepository.create({
     ...user,
     hashedPassword,
