+++ src/utils/auth.util.ts
@@ -1,6 +1,9 @@
 import bcrypt from "bcrypt";
 import { Client, Mover } from "@prisma/client";
 import { ConflictError } from "../types/errors";
+import { ErrorMessage } from "../constants/ErrorMessage";
+import authClientRepository from "../repositories/authClient.repository";
+import { ISignUpDataLocal } from "../types";
 
 // ✅ 비밀번호 해싱 함수
 export function hashPassword(password: string) {
@@ -21,4 +24,25 @@ export async function verifyPassword(inputPassword: string, savedPassword: strin
 
   const isValid = await bcrypt.compare(inputPassword, savedPassword);
   if (!isValid) throw new ConflictError("비밀번호를 잘못 입력하셨습니다.");
+}
+
+// ✅ 회원가입 유효성 검사
+export async function validateSignUpData(user: ISignUpDataLocal) {
+  const errors: Partial<Record<keyof ISignUpDataLocal, string>> = {};
+
+  const existingEmail = await authClientRepository.findByEmail(user.email);
+  const existingPhone = await authClientRepository.findByPhone(user.phone);
+
+  // 이미 사용 중인 정보
+  if (existingEmail) {
+    errors.email = ErrorMessage.ALREADY_EXIST_EMAIL;
+  }
+
+  if (existingPhone) {
+    errors.phone = ErrorMessage.ALREADY_EXIST_PHONE;
+  }
+
+  if (Object.keys(errors).length > 0) {
+    throw new ConflictError(`회원가입 실패: ${JSON.stringify(errors)}`);
+  }
 }
