+++ src/utils/authClient.util.ts
@@ -1,0 +1,2 @@
+// 나중에 파일 쓸 일 있을까 싶어 지우지는 않음.
+// 공용 함수는 auth.util.ts로 통합
